chore: change utils functions
