+++ library/lib/edges/EdgeProps.ts
@@ -1,10 +1,12 @@
 import { Edge, EdgeProps } from "@xyflow/react"
+import { IPoint } from "./Connection"
 
 export type CustomEdgeProps = {
   sourceRole: string | null
   sourceMultiplicity: string | null
   targetRole: string | null
   targetMultiplicity: string | null
+  points: IPoint[]
 }
 
 export type ExtendedEdgeProps = EdgeProps<Edge<CustomEdgeProps>> & {
