+++ library/lib/styles/app.css
@@ -87,16 +87,35 @@ body {
   background-color: rgb(99, 154, 242);
 }
 
-.edge-container:hover .edge-overlay {
+
+.edge-circle {
+  opacity: 0;
+  background-color: rgb(99, 154, 242);
+}
+
+.react-flow__edge:hover .edge-circle {
+  opacity: 0.4;
+  padding: 4px;
+  fill: #007bff; /* Blue color for handles */
+}
+
+.react-flow__edge:hover .edge-overlay {
   opacity: 0.4;
   padding: 4px;
   stroke: #007bff; /* Blue color for edge */
 }
 
-.edge-container:hover .edge-handle {
+
+.react-flow__edge.selected .edge-circle{
   opacity: 0.4;
   padding: 4px;
   fill: #007bff; /* Blue color for handles */
+}
+
+.react-flow__edge.selected .edge-overlay{
+  opacity: 0.4;
+  padding: 4px;
+  stroke: #007bff; /* Blue color for edge */
 }
 
 .react-flow__node:hover {
