+++ library/lib/utils/edgeUtils.ts
@@ -5,6 +5,7 @@ import {
   RHOMBUS_MARKER_PADDING,
   TRIANGLE_MARKER_PADDING,
 } from "@/constants"
+import { IPoint } from "@/edges/Connection";
 import { Position, Rect, XYPosition } from "@xyflow/react"
 /**
  * Adjusts the target coordinates based on the position and marker padding.
@@ -284,3 +285,125 @@ export function findClosestHandle(point: XYPosition, rect: Rect): string {
 
   return closest.label
 }
+// Utilss
+
+export function simplifySvgPath(path: string, decimals: number = 2): string {
+  const round = (num: number) => Number(num.toFixed(decimals))
+
+  const withSpaces = path.replace(/([MLQ])(?=[-0-9])/gi, "$1 ")
+  const cleaned = withSpaces.replace(/,/g, " ").trim()
+  const tokens = cleaned.split(/\s+/)
+  const outputTokens: string[] = []
+  let i = 0
+
+  while (i < tokens.length) {
+    const token = tokens[i].toUpperCase()
+    if (token === "M" || token === "L") {
+      const x = parseFloat(tokens[i + 1])
+      const y = parseFloat(tokens[i + 2])
+      if (!isNaN(x) && !isNaN(y)) {
+        outputTokens.push(token, round(x).toString(), round(y).toString())
+      }
+      i += 3
+    } else if (token === "Q") {
+      const cx = parseFloat(tokens[i + 1])
+      const cy = parseFloat(tokens[i + 2])
+      const ex = parseFloat(tokens[i + 3])
+      const ey = parseFloat(tokens[i + 4])
+      if (cx === ex && cy === ey) {
+        outputTokens.push("L", round(ex).toString(), round(ey).toString())
+      } else {
+        outputTokens.push(
+          "Q",
+          round(cx).toString(),
+          round(cy).toString(),
+          round(ex).toString(),
+          round(ey).toString()
+        )
+      }
+      i += 5
+    } else {
+      const x = parseFloat(tokens[i])
+      const y = parseFloat(tokens[i + 1])
+      if (!isNaN(x) && !isNaN(y)) {
+        outputTokens.push(round(x).toString(), round(y).toString())
+      }
+      i += 2
+    }
+  }
+  return outputTokens.join(" ")
+}
+
+export function simplifyPoints(points: IPoint[]): IPoint[] {
+  if (points.length < 3) return points
+  const result: IPoint[] = [points[0]]
+
+  for (let i = 1; i < points.length - 1; i++) {
+    const prev = result[result.length - 1]
+    const curr = points[i]
+    const next = points[i + 1]
+    if (prev.x === curr.x && curr.x === next.x) {
+      continue
+    }
+    if (prev.y === curr.y && curr.y === next.y) {
+      continue
+    }
+    result.push(curr)
+  }
+  result.push(points[points.length - 1])
+  return result
+}
+
+export function parseSvgPath(path: string): IPoint[] {
+  const tokens = path.replace(/,/g, " ").trim().split(/\s+/)
+  const points: IPoint[] = []
+  let i = 0
+  while (i < tokens.length) {
+    const token = tokens[i]
+    if (token === "M" || token === "L") {
+      const x = parseFloat(tokens[i + 1])
+      const y = parseFloat(tokens[i + 2])
+      if (!isNaN(x) && !isNaN(y)) {
+        points.push({ x, y })
+      }
+      i += 3
+    } else {
+      const x = parseFloat(tokens[i])
+      const y = parseFloat(tokens[i + 1])
+      if (!isNaN(x) && !isNaN(y)) {
+        points.push({ x, y })
+      }
+      i += 2
+    }
+  }
+  return simplifyPoints(points)
+}
+
+export function calculateInnerMidpoints(
+  points: IPoint[],
+  decimals: number = 2
+): IPoint[] {
+  const round = (num: number) => Number(num.toFixed(decimals))
+  const midpoints: IPoint[] = []
+  if (points.length < 4) return midpoints
+  for (let i = 1; i < points.length - 2; i++) {
+    const p1 = points[i]
+    const p2 = points[i + 1]
+    midpoints.push({ x: round((p1.x + p2.x) / 2), y: round((p1.y + p2.y) / 2) })
+  }
+  return midpoints
+}
+
+export function removeDuplicatePoints(points: IPoint[]): IPoint[] {
+  if (points.length === 0) return points
+  const filtered: IPoint[] = [points[0]]
+  for (let i = 1; i < points.length; i++) {
+    const prev = filtered[filtered.length - 1]
+    const current = points[i]
+    if (current.x !== prev.x || current.y !== prev.y) {
+      filtered.push(current)
+    }
+  }
+  return filtered
+}
+
